front: fix-bug: display rolling stock images
